// rivetkit-typescript/packages/rivetkit/src/common/bare/generated/inspector/v1.ts
export type PatchStateRequest = {
	readonly state: ArrayBuffer;
};

export function readPatchStateRequest(bc: bare.ByteCursor): PatchStateRequest {
	return {
		state: bare.readData(bc),
	};
}

export function writePatchStateRequest(
	bc: bare.ByteCursor,
	x: PatchStateRequest,
): void {
	bare.writeData(bc, x.state);
}

export type ActionRequest = {
	readonly id: uint;
	readonly name: string;
	readonly args: ArrayBuffer;
};

export function readActionRequest(bc: bare.ByteCursor): ActionRequest {
	return {
		id: bare.readUint(bc),
		name: bare.readString(bc),
		args: bare.readData(bc),
	};
}

export function writeActionRequest(
	bc: bare.ByteCursor,
	x: ActionRequest,
): void {
	bare.writeUint(bc, x.id);
	bare.writeString(bc, x.name);
	bare.writeData(bc, x.args);
}

export type StateRequest = {
	readonly id: uint;
};

export function readStateRequest(bc: bare.ByteCursor): StateRequest {
	return {
		id: bare.readUint(bc),
	};
}

export function writeStateRequest(bc: bare.ByteCursor, x: StateRequest): void {
	bare.writeUint(bc, x.id);
}

export type ConnectionsRequest = {
	readonly id: uint;
};

export function readConnectionsRequest(
	bc: bare.ByteCursor,
): ConnectionsRequest {
	return {
		id: bare.readUint(bc),
	};
}

export function writeConnectionsRequest(
	bc: bare.ByteCursor,
	x: ConnectionsRequest,
): void {
	bare.writeUint(bc, x.id);
}

export type EventsRequest = {
	readonly id: uint;
};

export function readEventsRequest(bc: bare.ByteCursor): EventsRequest {
	return {
		id: bare.readUint(bc),
	};
}

export function writeEventsRequest(
	bc: bare.ByteCursor,
	x: EventsRequest,
): void {
	bare.writeUint(bc, x.id);
}

export type ClearEventsRequest = {
	readonly id: uint;
};

export function readClearEventsRequest(
	bc: bare.ByteCursor,
): ClearEventsRequest {
	return {
		id: bare.readUint(bc),
	};
}

export function writeClearEventsRequest(
	bc: bare.ByteCursor,
	x: ClearEventsRequest,
): void {
	bare.writeUint(bc, x.id);
}

export type RpcsListRequest = {
	readonly id: uint;
};

export function readRpcsListRequest(bc: bare.ByteCursor): RpcsListRequest {
	return {
		id: bare.readUint(bc),
	};
}

export function writeRpcsListRequest(
	bc: bare.ByteCursor,
	x: RpcsListRequest,
): void {
	bare.writeUint(bc, x.id);
}

export type ToServerBody =
	| { readonly tag: "PatchStateRequest"; readonly val: PatchStateRequest }
	| { readonly tag: "StateRequest"; readonly val: StateRequest }
	| { readonly tag: "ConnectionsRequest"; readonly val: ConnectionsRequest }
	| { readonly tag: "ActionRequest"; readonly val: ActionRequest }
	| { readonly tag: "EventsRequest"; readonly val: EventsRequest }
	| { readonly tag: "ClearEventsRequest"; readonly val: ClearEventsRequest }
	| { readonly tag: "RpcsListRequest"; readonly val: RpcsListRequest };
// ...
export type ToServer = {
	readonly body: ToServerBody;
};